pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.6.0",
    "pydantic-ai-slim[logfire,mcp,openai]>=0.1.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.5",
]
